requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.7.1
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
click-completion==0.5.1
click-didyoumean==0.0.3
click-help-colors==0.5
colorama==0.4.1
comet-git-pure==0.19.11
comet-ml==2.0.5
configobj==5.0.6
cryptography==2.7
cycler==0.10.0
datadog==0.29.3
decorator==4.4.0
efficientnet==0.0.3
enum34==1.1.6
everett==1.0.2
future==0.17.1
gast==0.2.2
gradient==0.2.1
gradient-sdk==0.0.2
gradient-statsd==1.0.1
grpcio==1.20.0
h5py==2.9.0
hyperopt==0.1.2
idna==2.8
imageio==2.5.0
importlib-metadata==0.18
Jinja2==2.10.1
jsonschema==3.0.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mock==2.0.0
more-itertools==7.2.0
netifaces==0.10.9
networkx==2.3
numpy==1.16.3
nvidia-ml-py3==7.352.0
opencv-python==4.1.0.25
packaging==19.0
pandas==0.24.2
pbr==5.1.3
Pillow==6.0.0
pluggy==0.12.0
progressbar2==3.42.0
protobuf==3.7.1
py==1.8.0
pycparser==2.19
pymongo==3.8.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.15.3
pytest==5.0.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-utils==2.3.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
requests==2.22.0
requests-toolbelt==0.9.1
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
shellingham==1.3.1
six==1.12.0
termcolor==1.1.0
terminaltables==3.1.0
tqdm==4.32.1
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.2
wurlitzer==1.0.3
zipp==0.5.2
